+++ projects/tempo/BeatTemp.h
@@ -13,21 +13,21 @@ namespace vocalshaper {
 		~BeatTemp() = default;
 
 		struct BeatData final {
-			uint32_t xInBeat = 0;	//位置，以拍记
-			uint32_t xInBar = 0;	//位置，以小节记
-			uint8_t beat = 4;		//拍号
+			double xInBeat = 0;	//位置，以拍记
+			double xInBar = 0;	//位置，以小节记
+			uint8_t beat = 4;	//拍号
 		};
 
 		//由标签缓存刷新节拍缓存
 		void refresh();
 
 		//从缓存获取数据
-		uint8_t getBeatAtTime(uint32_t x) const;
-		uint32_t getBarAtTime(uint32_t x) const;
-		uint32_t getTimeAtBar(uint32_t bar) const;
+		uint8_t getBeatAtTime(double x) const;
+		double getBarAtTime(double x) const;
+		double getTimeAtBar(double bar) const;
 
-		int selectBy_x(uint32_t x) const;
-		int selectBy_bar(uint32_t bar) const;
+		int selectBy_x(double x) const;
+		int selectBy_bar(double bar) const;
 
 	private:
 		juce::Array<BeatData> list;
